// apps/frontend/src/app/dto/user.dto.ts
export class UserDTO {
  id?: string;
  name?: string;
  email?: string;
  avatarFileName?: string;
  birthDate?: string;
  description?: string;
  role?: string;
  sex?: string;
  location?: string;
  backgroundImgFileName?: string;
  trainingLevel?: string;
  trainingType?: string[];
  merits?: string;
  isReadyForTraining?: boolean;
  friends?: string[];
}